[od] Configure commander and gmenu2x resolution
Commander tries to detect the best resolution at startup.
This does not work correctly when started from an app that was running
in a different resolution, because the kernel always reports
the current resolution as the best video mode.

gmenu2x instead tries all resolutions from its hardcoded "supported"
list from highest to lowest resolution. There is no need to do that
because we already know the target resolution.